Refactor move TaskResult and TaskState models to schema module
Add a bit more documentation to models
